requirements.txt
# Dev environment
pip
autopep8

# App
flaskgoogle-generativeai
python-dotenv
python-dotenv
langchain-google-genai
pinecone
Flask-SQLAlchemy
gunicorn
mem0ai
chromadb
